requirements.txt
aiofiles~=22.1.0
aiosqlite~=0.17.0
apscheduler>=3.8,<4
hikari[speedups]==2.0.0.dev111
hikari-lightbulb>=2.2.1,<3
isodate~=0.6.1
psutil>=5.8,<6
pygount>=1.2,<2
python-dotenv~=0.21.0
rapidfuzz>=2.0.7,<3
scrapetube>=2.3,<3
uvloop~=0.17.0; os_name != "nt"
aiohttp~=3.8.3
pytz~=2022.4